created pages folder and moved proper files
