+++ src/App.js
@@ -2,9 +2,10 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import './styles/App.css';
 import Navbar from './components/Navbar';
-import Home from './components/Home.js'
-import About from './components/About';
-import Dashboard from './components/Dashboard';
+import Home from './pages/Home.js'
+import About from './pages/About';
+import Dashboard from './pages/Dashboard';
+import NoSuchPage from './pages/NoSuchPage';
 import Footer from './components/Footer';
 
 
@@ -17,6 +18,7 @@ function App() {
           <Route path="/" element={<Home />} />
           <Route path="/about" element={<About />} />
           <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="*" element={<NoSuchPage />} />
         </Routes>
       </div>
       <Footer />
