+++ src/pages/NoSuchPage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+
+const NoSuchPage = () => {
+    return(
+        <div id='nosuchpage'>
+            <h1 style={{textAlign: 'center'}}>Content: NoSuchPage</h1>
+            <h2 style={{textAlign: 'center', textDecoration: 'underline'}}>Error: No Page Found</h2>
+        </div>
+    );
+}
+
+export default NoSuchPage;
