+++ src/styles/App.css
@@ -111,6 +111,8 @@ a {
 /* --- END OF MOBILE HAMBURGER MENU (within navbar) --- */
 /* ===== END OF NAVBAR ===== */
 
+/* --- changes background image to smaller version --- */
+/*    divStyle.backgroundImage = `url(${BGimageMobile})`    */
 
 
 /* ===== START OF FOOTER ===== */
